requirements.txt
weaviate-client
uvicorn
fastapi
